requirements.txt
# MAIFS - Multi-Agent Image Forensic System
# Requirements

# Core Dependencies
numpy>=1.21.0
scipy>=1.7.0
pillow>=9.0.0

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0

# Optional: LangChain Integration
# langchain>=0.1.0
# langchain-openai>=0.0.5
# langchain-anthropic>=0.1.0

# Optional: Local LLM
# llama-cpp-python>=0.2.0
# transformers>=4.35.0

# Web UI
gradio>=4.0.0

# Image Processing
opencv-python>=4.5.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocstrings[python]>=0.24.0
